fix types of variables

+++ src/main/java/nflx/rozhnov/transactionservice/dto/response/TransactionRs.java
@@ -2,6 +2,7 @@ package nflx.rozhnov.transactionservice.dto.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -13,7 +14,7 @@ import java.util.UUID;
 public class TransactionRs {
     private UUID transactionId;
     private Date timestamp;
-    private long fromAccount;
-    private long toAccount;
-    private double amount;
+    private Long fromAccount;
+    private Long toAccount;
+    private BigDecimal amount;
 }
